-Merged Put and Get DTO for User and UserPreferences
