+++ src/main/java/ch/uzh/ifi/hase/soprafs22/rest/mapper/DTOMapper.java
@@ -34,23 +34,16 @@ public interface DTOMapper {
   @Mapping(source = "status", target = "status")
   @Mapping(source = "birthday", target = "birthday")
   @Mapping(source = "gender", target = "gender")
+  @Mapping(source = "minAge", target = "minAge")
+  @Mapping(source = "maxAge", target = "maxAge")
+  @Mapping(source = "genderPreferences", target = "genderPreferences")
   UserGetDTO convertEntityToUserGetDTO(User user);
-
-
-  @Mapping(source = "minAge", target = "minAge")
-  @Mapping(source = "maxAge", target = "maxAge")
-  @Mapping(source = "genderPreferences", target = "genderPreferences")
-  UserGetPreferencesDTO convertEntityToUserPreferencesGetDTO(User user);
-
-  @Mapping(source = "id", target = "id")
-  @Mapping(source = "minAge", target = "minAge")
-  @Mapping(source = "maxAge", target = "maxAge")
-  @Mapping(source = "genderPreferences", target = "genderPreferences")
-  User convertUserPreferencesPutDTOtoEntity(UserPreferencesPutDTO userPutPreferencesDTO);
 
   @Mapping(source = "id", target = "id")
   @Mapping(source = "username", target = "username")
-  @Mapping(source = "birthday", target = "birthday")
+  @Mapping(source = "minAge", target = "minAge")
+  @Mapping(source = "maxAge", target = "maxAge")
+  @Mapping(source = "genderPreferences", target = "genderPreferences")
   User convertUserPutDTOtoEntity(UserPutDTO userPutDTO);
 
   @Mapping(source = "id", target = "id")
